requirements.txt
httpx>=0.24.0
beautifulsoup4>=4.12.0
python-telegram-bot>=20.0
